backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
Celery==5.3.0
redis==5.0.0
